requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
anyio==4.7.0
APScheduler==3.6.3
attrs==24.3.0
cachetools==4.2.2
certifi==2024.12.14
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
propcache==0.2.1
python-telegram-bot==21.9
pytz==2024.2
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.18.3
